--- src/main/java/com/revature/devkrogergrocerylistapp/Test.java
@@ -1,2 +1,0 @@
-package com.revature.devkrogergrocerylistapp;public class Test {
-}
