+++ src/main/java/com/revature/vanqapp/Test.java
@@ -1,0 +1,4 @@
+package com.revature.vanqapp;
+
+public class Test {
+}
